Create a screen that allows a user to manage a list of Customers, the Customer should have a Name and a Type.
Name is required and must have a minimum length of 5 characters
Type is required and should be a select box of either "Large" or "Small".
There should be a form to add a new customer (like on the jobs page).
There should be a list to see all the customers (like on the jobs page).
There should be a link on the list to open the customer record (like on the jobs page).