The jobs page needs to be extended to allow assigning a job to a customer.
When creating a job the user should be able to pick a customer from the dropdown.
Selecting a customer should be required for creating a job.
The user should be able to see assigned customer in the jobs list.
For any existing jobs that were not assigned to a customer it should display "Unknown" in the list.
When the user opens the job details from the list this screen should include information about the Customer - Name and Type.